+++ CHANGELOG.md
@@ -6,3 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Bedrock Guardrails sample
+
+<!-- markdownlint-disable MD024 -->
+## [1.0.0]
+
+### Added
+
+- Document Explorer sample
+- Bedrock Agent sample